Remove MimeTypeManager
