+++ core/impl/src/main/java/cloud/piranha/core/impl/DefaultWebApplicationManager.java
@@ -35,7 +35,6 @@ import cloud.piranha.core.api.HttpSessionManager;
 import cloud.piranha.core.api.JspManager;
 import cloud.piranha.core.api.LocaleEncodingManager;
 import cloud.piranha.core.api.LoggingManager;
-import cloud.piranha.core.api.MimeTypeManager;
 import cloud.piranha.core.api.MultiPartManager;
 import cloud.piranha.core.api.ObjectInstanceManager;
 import cloud.piranha.core.api.SecurityManager;
@@ -92,11 +91,6 @@ public class DefaultWebApplicationManager implements WebApplicationManager {
      * Stores the logging manager.
      */
     protected LoggingManager loggingManager = new DefaultLoggingManager();
-    
-    /**
-     * Stores the mime-type manager.
-     */
-    protected MimeTypeManager mimeTypeManager = new DefaultMimeTypeManager();
 
     /**
      * Stores the multi-part manager.
@@ -179,11 +173,6 @@ public class DefaultWebApplicationManager implements WebApplicationManager {
     }
 
     @Override
-    public MimeTypeManager getMimeTypeManager() {
-        return mimeTypeManager;
-    }
-
-    @Override
     public MultiPartManager getMultiPartManager() {
         return multiPartManager;
     }
@@ -256,11 +245,6 @@ public class DefaultWebApplicationManager implements WebApplicationManager {
     @Override
     public void setLoggingManager(LoggingManager loggingManager) {
         this.loggingManager = loggingManager;
-    }
-
-    @Override
-    public void setMimeTypeManager(MimeTypeManager mimeTypeManager) {
-        this.mimeTypeManager = mimeTypeManager;
     }
 
     @Override
